Move tkn-hub into api for easier vendoring into tkn cmd
With cli being a separate module, adding hub/cli into `tkn` is
quite difficult because of the `require`

```
require api => ../api
```

Thus we are moving `tkn-hub` cli into `api/cmd` so that `tkn`
just needs to add `api` as dependency and depend on `hub/api/pkg/cli/*`
