+++ api/pkg/cli/cmd/search/testdata/TestSearch_JSONFormat.golden
