+++ api/pkg/cli/cmd/search/testdata/TestSearch_TableFormat.golden
