+++ api/pkg/cli/hub/search_test.go
